Define MapTile as derived from BindTile
