+++ core/src/main/scala/soda/tiles/fairness/tile/derived/bind/SigmaTile.soda
