--- core/src/main/scala/soda/tiles/fairness/tile/derived/map/Package.soda
@@ -1,14 +1,0 @@
-package soda.tiles.fairness.tile.derived.map
-
-/*
- * This package contains classes to model the tiles.
- */
-
-import
-  soda.tiles.fairness.tile.primitive.MapTile
-  soda.tiles.fairness.tool.Agent
-  soda.tiles.fairness.tool.Measure
-  soda.tiles.fairness.tool.TileMessage
-  soda.tiles.fairness.tool.TileMessageBuilder
-  soda.tiles.fairness.tool.TilePair
-  soda.tiles.fairness.tool.TileTriple
