+++ examples/src/main/scala/soda/tiles/fairness/example/pipeline/complementtheilindex/Package.soda
@@ -7,12 +7,12 @@ package soda.tiles.fairness.example.pipeline.complementtheilindex
 import
   soda.tiles.fairness.tile.composite.AccumulatesTile
   soda.tiles.fairness.tile.constant.AllAgentTile
+  soda.tiles.fairness.tile.derived.bind.MapTile
   soda.tiles.fairness.tile.derived.fold.LengthTile
   soda.tiles.fairness.tile.derived.fold.SumPhiTile
   soda.tiles.fairness.tile.derived.fold.SumPhiNumberTile
   soda.tiles.fairness.tile.primitive.ApplyTile
   soda.tiles.fairness.tile.primitive.CrossTile
-  soda.tiles.fairness.tile.primitive.MapTile
   soda.tiles.fairness.tile.primitive.TuplingPairTile
   soda.tiles.fairness.tool.Agent
   soda.tiles.fairness.tool.Measure
