+++ examples/src/main/scala/soda/tiles/fairness/example/pipeline/equity/Package.soda
@@ -8,7 +8,7 @@ import
   soda.tiles.fairness.tile.composite.AccumulatesTile
   soda.tiles.fairness.tile.composite.AllAtLeastTile
   soda.tiles.fairness.tile.constant.AllAgentTile
-  soda.tiles.fairness.tile.derived.map.NeedsTile
+  soda.tiles.fairness.tile.derived.bind.NeedsTile
   soda.tiles.fairness.tool.Agent
   soda.tiles.fairness.tool.Measure
   soda.tiles.fairness.tool.Number
